requirements.txt
# Core dependencies
langgraph>=1.0.4,<2.0.0
langchain>=1.1.0,<2.0.0
langchain-openai>=1.1.0,<2.0.0
langchain-anthropic>=1.2.0,<2.0.0

# Document processing
docling>=2.64.0,<3.0.0
easyocr>=1.7.2,<2.0.0
pytesseract>=0.3.13,<0.4.0

# API
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.32.0,<1.0.0
python-multipart>=0.0.12,<1.0.0

# Utilities
pydantic>=2.12.5,<3.0.0
imap-tools>=1.11.0,<2.0.0
python-dotenv>=1.2.1,<2.0.0
httpx>=0.27.2,<1.0.0

# Dev
pytest>=9.0.1,<10.0.0
